// action/src/schemas/stoatConfigSchema.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface StoatConfigSchema {
  version: number;
  enabled?: boolean;
  comment_template_file?: string;
  plugins?: {
    static_hosting?: StaticHostingPluginMap;
    json?: JsonPluginMap;
    image_diff?: ImageDiffPluginMap;
    job_runtime?: JobRuntimePlugin;
    workflow_dispatch?: WorkflowDispatchPluginMap;
    [k: string]: unknown;
  };
  [k: string]: unknown;
}
export interface StaticHostingPluginMap {
  [k: string]: StaticHostingPlugin;
}
export interface StaticHostingPlugin {
  metadata?: {
    [k: string]: unknown;
  };
  path: string;
  file_viewer?: boolean;
  [k: string]: unknown;
}
export interface JsonPluginMap {
  [k: string]: JsonPlugin;
}
export interface JsonPlugin {
  metadata?: {
    [k: string]: unknown;
  };
  path: string;
  [k: string]: unknown;
}
export interface ImageDiffPluginMap {
  [k: string]: ImageDiffPlugin;
}
export interface ImageDiffPlugin {
  metadata?: {
    [k: string]: unknown;
  };
  image: string;
  baseline: string;
  [k: string]: unknown;
}
export interface JobRuntimePlugin {
  enabled: boolean;
  tracking?: boolean;
  chart?: {
    width?: number;
    height?: number;
    [k: string]: unknown;
  };
  [k: string]: unknown;
}
export interface WorkflowDispatchPluginMap {
  [k: string]: WorkflowDispatchPlugin;
}
export interface WorkflowDispatchPlugin {
  filename: string;
  /**
   * All workflow related information will be persisted under "plugins.workflow_dispatch.<task-id>.<identifier>". This is useful if the same workflow can be triggered for different purposes (e.g. deployment for multiple environments).
   */
  identifier?: string;
  [k: string]: unknown;
}